Add deploy step
This required a revamp of the functions to mostly use resource IDs,
instead of relying on names. The way we retrieve the IDs is a bit ugly
and should be looked at later. Currently, the script will succeed even
if the compile/deploy steps fail, because they are asynchronously and
fail silently. This should also be imporved.