Skip updating unchanged deployments
The only thing that can currently change with deployments is the set
of tags. To further reduce the number of API-GW roundtrips, we parse
the JSON to get the current set of tags and compare it to the new
set of tags.

To do this, we also generalize the list generation procedure that already
exists in a few places. Sorting the lists makes output nice, and
enables exact string comparisons between  two lists.